Port iosp.Layout to array. Remove old ma2 code from H5iosp. enums have ArrayType.primitive classes match others.
