+++ cdm-core/src/main/java/ucar/nc2/iosp/LayoutBBTiled.java
@@ -4,8 +4,8 @@
  */
 package ucar.nc2.iosp;
 
-import ucar.ma2.InvalidRangeException;
-import ucar.ma2.Section;
+import ucar.array.InvalidRangeException;
+import ucar.array.Section;
 import java.io.IOException;
 import java.nio.*;
 
@@ -15,7 +15,6 @@ import java.nio.*;
  * "Tiled" means that all chunks are assumed to be equal size.
  * Chunks do not necessarily cover the array, missing data is possible.
  * Used by HDF4 and HDF5.
- * TODO will switch to ucar.array.* in ver8.
  */
 public class LayoutBBTiled implements LayoutBB {
   private static final boolean debug = false, debugIntersection = false;
