+++ cdm-core/src/main/java/ucar/nc2/iosp/LayoutTiled.java
@@ -4,9 +4,9 @@
  */
 package ucar.nc2.iosp;
 
-import ucar.ma2.InvalidRangeException;
-import ucar.ma2.Range;
-import ucar.ma2.Section;
+import ucar.array.InvalidRangeException;
+import ucar.array.Range;
+import ucar.array.Section;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -17,7 +17,6 @@ import java.util.List;
  * Chunks have an offset into the complete array.
  * Chunks do not necessarily cover the array, missing data is possible.
  * Used by HDF4 and HDF5.
- * TODO will switch to ucar.array.* in ver8.
  */
 public class LayoutTiled implements Layout {
   private static final boolean debug = false, debugNext = false;
