+++ netcdf4/src/test/java/ucar/nc2/jni/netcdf/TestReadNc4Compare.java
@@ -11,7 +11,6 @@ import org.junit.runners.Parameterized;
 import ucar.nc2.NetcdfFile;
 import ucar.nc2.NetcdfFiles;
 import ucar.nc2.internal.util.CompareArrayToArray;
-import ucar.nc2.internal.util.CompareArrayToMa2;
 import ucar.nc2.internal.util.CompareNetcdf2;
 import ucar.unidata.util.test.TestDir;
 import ucar.unidata.util.test.category.NeedsCdmUnitTest;
@@ -28,7 +27,7 @@ import static com.google.common.truth.Truth.assertThat;
 @RunWith(Parameterized.class)
 @Category(NeedsCdmUnitTest.class)
 public class TestReadNc4Compare {
-  public static String iospOrg = "ucar.nc2.jni.netcdf.Nc4readerOrg";
+  public static String iospOrg = "ucar.nc2.jni.netcdf.Nc4reader";
   public static String iospArray = "ucar.nc2.jni.netcdf.Nc4reader";
 
   @Parameterized.Parameters(name = "{0}")
@@ -64,7 +63,7 @@ public class TestReadNc4Compare {
         NetcdfFile arrayFile = NetcdfFiles.open(filename, iospArray, -1, null, null)) {
       System.out.println("Test NetcdfFile: " + arrayFile.getLocation());
 
-      boolean ok = CompareArrayToMa2.compareFiles(ma2File, arrayFile);
+      boolean ok = CompareArrayToArray.compareFiles(ma2File, arrayFile);
       assertThat(ok).isTrue();
     }
   }
