+++ toolsui/uicdm/src/main/java/ucar/ui/op/Hdf5NewDataTable.java
@@ -13,7 +13,6 @@ import ucar.nc2.Variable;
 import ucar.nc2.internal.iosp.hdf5.H5diagNew;
 import ucar.nc2.internal.iosp.hdf5.H5header;
 import ucar.nc2.internal.iosp.hdf5.H5iosp;
-import ucar.nc2.internal.iosp.hdf5.H5iospArrays;
 import ucar.ui.widget.BAMutil;
 import ucar.ui.widget.PopupMenu;
 import ucar.ui.widget.TextHistoryPane;
@@ -142,7 +141,7 @@ public class Hdf5NewDataTable extends Hdf5DataTable {
     this.location = raf.getLocation();
     List<VarBean> beanList = new ArrayList<>();
 
-    iosp = new H5iospArrays();
+    iosp = new H5iosp();
     try {
       ncfile = NetcdfFiles.build(iosp, raf, raf.getLocation(), null);
     } catch (Throwable t) {
